+++ clients/ios/Other Sources/InAppSettingsKit/Models/IASKSettingsStore.h
@@ -29,7 +29,7 @@
 - (BOOL)boolForKey:(NSString*)key;
 - (float)floatForKey:(NSString*)key;
 - (double)doubleForKey:(NSString*)key;
-- (int)integerForKey:(NSString*)key;
+- (NSInteger)integerForKey:(NSString*)key;
 - (id)objectForKey:(NSString*)key;
 - (BOOL)synchronize; // Write settings to a permanant storage. Returns YES on success, NO otherwise
 @end
